+++ EasyPost.NetFramework/Report.cs
@@ -65,17 +65,17 @@ namespace EasyPost
         ///     All invalid keys will be ignored.
         /// </param>
         /// <param name="type">The type of report, e.g. "shipment", "tracker", "payment_log", etc.</param>
-        /// <returns>Instance of EasyPost.ScanForm.</returns>
-        public static ReportList All(string type, Dictionary<string, object> parameters = null)
+        /// <returns>An EasyPost.ReportCollection instance.</returns>
+        public static ReportCollection All(string type, Dictionary<string, object> parameters = null)
         {
             Request request = new Request("reports/{type}");
             request.AddUrlSegment("type", type);
             request.AddQueryString(parameters ?? new Dictionary<string, object>());
 
-            ReportList reportList = request.Execute<ReportList>();
-            reportList.filters = parameters;
-            reportList.type = type;
-            return reportList;
+            ReportCollection reportCollection = request.Execute<ReportCollection>();
+            reportCollection.filters = parameters;
+            reportCollection.type = type;
+            return reportCollection;
         }