+++ EasyPost.NetFramework/ShipmentCollection.cs
@@ -4,7 +4,7 @@ using Newtonsoft.Json;
 
 namespace EasyPost
 {
-    public class ShipmentList : Resource
+    public class ShipmentCollection : Resource
     {
         [JsonProperty("filters")]
         public Dictionary<string, object> filters { get; set; }
@@ -16,8 +16,8 @@ namespace EasyPost
         /// <summary>
         ///     Get the next page of shipments based on the original parameters passed to Shipment.All().
         /// </summary>
-        /// <returns>A new EasyPost.ShipmentList instance.</returns>
-        public ShipmentList Next()
+        /// <returns>An EasyPost.ShipmentCollection instance.</returns>
+        public ShipmentCollection Next()
         {
             filters = filters ?? new Dictionary<string, object>();
             filters["before_id"] = shipments.Last().id;